Backend/requirements.txt
requests
python-dotenv
sentence-transformers
faiss-cpu